requirements.txt
requests
pandas
geopandas
fastapi
uvicorn
python-dotenv
anthropic
shapely
